Move end-routes from server.js to app.js
